requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
hypothesis>=6.92.0

# Utilities
python-dateutil>=2.8.2

# CLI UI
rich>=13.0.0
